requirements.txt
-e git+https://github.com/lvermue/PyKEEN.git@dev2#egg=pykeen
alembic==1.0.11
asn1crypto==0.24.0
astroid==2.2.5
bayesian-optimization==1.0.1
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click-default-group==1.2.1
click==7.0
cloudpickle==1.2.1
configparser==3.7.4
coverage==4.5.4
cryptography==2.7
cycler==0.10.0
databricks-cli==0.8.7
dataclasses==0.6
docker==4.0.2
entrypoints==0.3
flask==1.1.1
future==0.17.1
gitdb2==2.0.5
gitpython==2.1.13
gorilla==0.3.0
gunicorn==19.9.0
idna==2.8
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
mako==1.0.14
markupsafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mlflow==1.1.0
numpy==1.17.0
pandas==0.25.0
paramiko==2.6.0
pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.0
pycparser==2.19
pylint==2.3.1
pynacl==1.3.0
pyparsing==2.4.1.1
pysftp==0.2.9
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
pyyaml==5.1.1
querystring-parser==1.2.4
rdflib==4.2.2
requests==2.22.0
scikit-learn==0.19.1
scipy==1.3.0
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
sparqlwrapper==1.8.4
sqlalchemy==1.3.6
sqlparse==0.3.0
tabulate==0.8.3
torch==1.1.0
torchvision==0.2.1
tqdm==4.32.2
typed-ast==1.4.0 ; implementation_name == 'cpython'
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
werkzeug==0.15.5
wrapt==1.11.2
